Dev_v2/voicenote/cloud-run/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Google Cloud
google-cloud-firestore>=2.11.1
google-cloud-storage>=2.10.0
google-cloud-tasks>=2.14.1
google-cloud-logging>=3.8.0

# 音声処理
pyannote.audio>=3.1.1
torch>=2.0.1
torchaudio>=2.0.2
librosa>=0.10.1
noisereduce>=3.0.0
pydub>=0.25.1
scipy>=1.11.1
numpy>=1.24.3
soundfile>=0.12.1

# 音声認識API
openai>=1.3.0
azure-cognitiveservices-speech>=1.31.0
google-cloud-speech>=2.21.0
assemblyai>=0.17.0
deepgram-sdk>=3.0.0

# LLM API
anthropic>=0.7.0
google-generativeai>=0.3.0

# ユーティリティ
python-dotenv>=1.0.0
pydantic>=2.4.0
aiofiles>=23.2.1
httpx>=0.25.0
tenacity>=8.2.3

# Hugging Face
huggingface-hub>=0.17.0
transformers>=4.35.0

# 機械学習
scikit-learn>=1.3.0
joblib>=1.3.2